Move sponsored broadcast into transaction.py, make wallet_address a callback contract
